Add files via upload
This version of the database contains 4 tables, a trigger and a function

The trigger  checks to see if any invoices have been made of a particular ID and has 2 reactions depending on the result
The procedure is what performs the selling of items and updates the tables effected based on the quantity given (Let it be known that this procedure does not have error handling in case the user inputs an invalid amount

Foreign keys have been established in sales_details, referencing sales_summary and the products table. Once a record from sales_summary or products has been removed, all records with common values in the sales_details table are removed.